Rename ConnectedDutyStation
